QML macros: suppress Wredundant-decls
When using multiple QML type registration macros, one might end up with
duplicated friend declarations. This is fine as far as C++ is concerned,
but triggers gcc's redundant-decls warning.

Given that we require the friend declaration, and would like to keep it
in the macros that actually need them (instead of putting them into the
macros which are used in all types exposed to QML), we opt for
suppressing the warning.

Pick-to: 6.5 6.7
Fixes: QTBUG-124934
Change-Id: I4f43afddfefb6e546ddddbdc01435222398a06be
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>