Scale VectorImage contents to match requested size
If the user sets a size on the VectorImage, then we should adapt
the contents to this, like we do with Image and Shape. This
also introduces a fillMode property that matches the ones in
Image and Shape and allows you to specify how to handle it
when the requested size differs from the implicit size.

Fixes: QTBUG-122696
Change-Id: Ic68e1acd0537d97216d5a4b5df6adeaa15e8e258
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>