QQuickPopupWindow: forward pointer events to the menu under the mouse
We want to handle menus special compared to other popups.
For menus, we want all open parent menus and sub menus that
belong together to almost act as a single popup WRT event
delivery. This will allow the user to hover and highlight
MenuItems inside all of them, and not just the leaf menu.

This patch will therefore forward all relevant pointer events
that goes to a leaf menu to the actual menu, or menu bar,
under the mouse. But note that this forwarding will happen
in parallel with normal event delivery (we don't eat the
events), as we don't want to break the delivery to e.g
grabbers, and other DA logic in general.

Change-Id: If87756e59c08ff73cb75a21fafab6cf39e4c94b6
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>