Force active focus to target item in setFocusToTargetItem
It is not enough to call setFocus on the target item, as there
may be other focus scopes currently holding the active focus.
Call forceActiveFocus instead.

Also, if the target item is the First or Last target, we want
to explicitly set the focus to the first/last item of the window
that can receive focus. If that item is a FocusScope we need to
clear its previous subFocusItem.

Task-number: QTBUG-121789
Change-Id: I99e5da5727dfd99c76bf2fc0baf7e9b8330f7c4d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>