documentationhints: introduce help manager
HelpManager class is responsible for initializing the help engine
plugin and perform the operations using the plugin. documentationForItem()
is invoked by the hover handler and returns the requested documentation
info.

For now, HelpManager greedily registers all the qch files it finds in
the given base documentation path. In the next iteration, the better
approach might be only registering the imported modules documentations.

It has utility functions that finds the hovered keyword and tries to
fetch documentation for it.

style: Also remove superfluous blanklines in unrelated positions

Pick-to: 6.8
Task-number: QTBUG-120016
Change-Id: I151bc3dd1e224ccf51e66f10723c721ae622fae1
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>