Gallery example: add colorScheme selection
Works with Fusion, iOS, and FluentWin3 styles.

Should also work with macOS and ideally Basic, but those styles
do not respect the global palette yet. So disable the row with the
color scheme selection check boxes unless we know it works.

The color scheme setting is not stored across sessions, and is
instead applied "live".

Task-number: QTBUG-124490
Change-Id: If6bec56884e6dc9093d541fc87800596aa1b9ca3
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>