Support drag and drop section reordering in the table view
The new drag handler (QQuickTableViewDragHandler) in the header
view allows the user to drag and drop row or column at the
required index. At any point in time, there can be only one active
handler, and depending on the type of operation, the corresponding
handler consumes the pointer events. The user can customize the
delegate using the new property 'containsDrag' when any
corresponding drop section is activated while dragging a column or
row.

[ChangeLog][QtQuick][TableView] Added new property 'containsDrag'
for the delegate, and it is set when a user drags a column or row
on top of another delegate. This property is applicable only for
HorizontalHeaderView and VerticalHeaderView.

Task-number: QTBUG-100440
Change-Id: I52b1079559192f406e295aa34ca1d86acef61a6d
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>