Add and use QQPopupPrivate::handleReleaseWithoutGrab for menus
QQuickDeliveryAgent doesn't allow an item to see a pointer release event
unless the item has already grabbed; that's usually done on press.
But during the drag-press-release gesture, we don't know which menu item
the user will choose until the release. QQuickMenu isn't doing all the
event handling on its own, and thus isn't the grabber: QQuickMenuItem
inherits QQuickAbstractButton, and if that "button" didn't get grabbed,
it's not going to react to the release. So we need a workaround.

Given that the popup window did not see the mouse/touch/tablet press
but just some moves and a release, there will not be a grabber within
the popup window. In that case, on release, call virtual
QQuickPopupPrivate::handleReleaseWithoutGrab(), which is overridden in
QQuickMenuPrivate to programmatically "click" the currently-hovered menu
item. Then propagation continues, to allow the component that opened
the menu to also see the release, and exit from its "pressed" state.

Task-number: QTBUG-68080
Change-Id: I3280ffe722560726003522bb33e5152302cb8cbc
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>