SplitView: fix resized item jumping in size when an item is hidden
The logic that handles the item to the left of the handle being resized
incorrectly used a hidden handle as the leftEdge. This would cause the
item to get more space than it should because it thought it had more
space to consume.

Before the fix, leftEdge is 10 and accumulated is 0, so:

    //          390         400    0             10
    const qreal rightStop = size - accumulated - pressedHandleSize;

    //    110             10        110
    qreal leftStop = qMax(leftEdge, pressedHandlePos);

    //          110                   110       110               390
    const qreal newHandlePos = qBound(leftStop, pressedHandlePos, rightStop);

    //          100           110            10
    const qreal newItemSize = newHandlePos - leftEdge;

After the fix, leftEdge is 60 and accumulated is 0, so:

    //          390         400    0             10
    const qreal rightStop = size - accumulated - pressedHandleSize;

    //    110             60        110
    qreal leftStop = qMax(leftEdge, pressedHandlePos);

    //          110                   110       110               390
    const qreal newHandlePos = qBound(leftStop, pressedHandlePos, rightStop);

    //          50            110            60
    const qreal newItemSize = newHandlePos - leftEdge;

Fixes: QTBUG-125049
Change-Id: I3cdf57693c71c1f367780eda0ff6794fea02d546
Reviewed-by: Kaj Grönholm <kaj.gronholm@qt.io>
(cherry picked from commit e11637d28f33d5e4c4d1d09fd23ef4364151d625)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 955f2e032dcf0b0f144fcde2f76525c3b366517c)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>