QQuickMenuBar: keep explicit track of open menus
After calling QQuickMenu::popup(pos), QQuickMenu::isVisible() will
not be true until the menu is actually visible on screen.
Even if this delay is sometimes only milliseconds, is enough to
fool QQuickMenuBar to think that the current menu is not supposed
to be open, if you quickly hover another MenuBarItem while the current
MenuBarItem is pending to be open. The result is that we "forget" to
close the current menu, and open the new one, when this happens.

This patch will therefore instead make sure that we keep explicit track
of whether the current MenuBarItem is supposed to be open or not, instead
of relying too much on QQuickMenu::isVisible().

Change-Id: Ie6df19f7ded71ab49a655e6997e79ee4f77ede13
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>