QQuickMenuBar: don't deactivate a hovered MenuBarItem
When a Menu is closed, we need to check if we should then also
deactivate the MenuBarItem that owns it. But we should only do
so if the MenuBarItem is no longer hovered, and its also the
active MenuBarItem.  The latter will not be the case when you
hover from one open MenuBarItem to the next, as in that case, it
will be the previous MenuBarItem that closes, and not the
current one.

This patch will ensure we follow the abovestanding logic.

Change-Id: I3ff7453e82907bdb33abebec2e808aeb2a9c0874
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>