quick_test_main_with_setup: only activate windows where supported
Otherwise we get warnings before initTestCase, meaning not even
failOnWarning can catch them:

"qWaitForWindowActive was called on a platform that doesn't support
window activation. This means there is an error in the test and it
should either check for the WindowActivation platform capability before
calling qWaitForWindowActivate, use qWaitForWindowExposed instead, or
skip the test. Falling back to qWaitForWindowExposed."

Task-number: QTBUG-124230
Pick-to: 6.5 6.7
Change-Id: I8111928917b94d51ac9b94f470f5837a98f44a71
Reviewed-by: Doris Verria <doris.verria@qt.io>