tst_qquickmenubar: disable popup windows until 100% ready
Many of the MenuBar tests will currently fail if using popup
windows. And some cannot be expected to work until event
forwarding and key navigation for popups is properly implemented.
At the same time, implementing the mentioned functionality depends
on this patch being merged first. So disable some of the test for now.

Change-Id: I8eedf08ca47e89b35eeb43206268c8ce7f582ef2
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>