Call polish after setting HAlign in Qml TextEdit
Qml TextEdit does not call polish() in the setter for horizontal alignment.

The test that tests horizontal alignment for TextEdit is broken,
creating a Text item instead of a TextEdit and therefore calling the
wrong setter.

This test also has a false positive as a result of an unnecessary
subtraction.

Fixes: QTBUG-56921
Pick-to: 5.15 6.2 6.5 6.7
Change-Id: I21df3c4f403980fb5512abfd5e7e32a064cd60d7
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>