qmlls: implement commitToBase workflow for lazy files
It seems that the current lazy QmlFile implementation does not take into
account the commitToBase workflow used in qmlls, which leads to crashes
and weird behaviors.

Change the logic in QDeferredFactory<QQmlJSScope> to allow a custom
TypeReader: QQmlJSTypeReader always reads from disk which is bad
because qmlls operates most of the time on in-memory files and
introduces an unnecessary IO operation, and therefore a new TypeReader
is introduced in this commit for files loaded in the Dom.

Populate lazy QmlFiles into the environment that they were loaded into
with the new TypeReader before commitToBase() calls, and into the base
environment after the commitToBase() call. This fixes the crashes during
completion where a heap-use-after-free happened when a lazy file tried
to populate itself after a commitToBase call using a destructed
temporary DomEnvironment.

Add a resetFactory() method to QDeferredSharedPtr to force
recompute a QQmlJSScope when a file was changed without having to
recreate all the qqmljsscopes for the other files, and use it also
during commitToBase() calls to update the environment needed for the
lazy loading. If the previous factory was used already, then
the qqmljsscope is eagerly populated to avoid stale data.

Also make sure to update the loadPaths for the build directory tests
when getting a SemanticAnalysis struct from the base environment.

Recreate the "completion crash" scenario from QTBUG-124799 in
tst_qmldomitem. Add tests to make sure that files are correctly
populated before and after commitToBase() calls.

Also add a test in tst_qmlls_qqmlcodemodel to make sure that files are
correctly lazy loaded in qqmlcodemodel, and simplify
qqmlcompletionsupport's completions method by early exiting.

Make populateFromQmlFile populate the logger in the same way as the
QQmlJSTypeReader does.

Fixes: QTBUG-124799
Fixes: QTBUG-125388
Change-Id: Id8bd7a1cbe60ba961eadb37b121f8f0670b863f4
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>