Load a QJsonArray property as a Sequence
When a property coming from C++ is used in a QML file, the engine might
perform some internal conversion to handle the type.

When a `QJsonArray` property is loaded, currently, it will be converted
to an internal `JsonObject` representation containing an engine-owned
array.

When the conversion is done, the data from the `QJsonArray` is copied as
a part of this process.

The performed copy is expected to act like a Javascript array.
Nonetheless, due to the way the conversion is implemented, using the
property will perform the requested work on the copy without writing
back to the original property.

This in turn prevents certain operation from working as expected, in
particular the mutable methods on the array prototype.

For example, if `jsonArray` is a `QJsonArray` property with a C++
provenance, the following will not register the push on the original
property, so that it will be lost the following time the property is
accessed:

```
...
jsonArray.push(4)
...
```

As a `QJsonArray` property should behave as a Javascript array when used
from QML, a `QJsonArray` property will now be converted to a `Sequence`,
an internal representation that behaves like a Javascript array but acts as a
reference object with property write-backs, to allow in-place mutations
to be correctly reflected on the original property.

Currently, a `QJsonArray` `Sequence` call to the sort method will not
correctly work due to the way the sort method is implemented.
This is currently left as non-working and will be fixed separately.

A non-exhaustive test-case was added to check that a C++ loaded
`QJsonProperty` acts as a Javascript array with regards to the Array
prototype methods, taking into consideration the inability of the sort
method to currently work.

Task-number: QTBUG-111015
Change-Id: Iec48fe4cba9adb9b794e5a568985b86b8da4556c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>