Add a wrapper builtin for QJsonArray
Calling `qmllint` can produce warning for usages of the `QJsonArray`
type, due to it not being found.

Nonetheless, the engine special handles the type, so that `qmllint`
should not warn about it.

Make `QJsonArray` a builtin type so that calls to `qmllint` will import
and recognize the type.

Add a test-case to ensure that the type is now recognized.

Task-number: QTBUG-111015
Change-Id: I51744968a0316ccb016a4404326b74fa419037bd
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>