Android: Add possibility to mark QML files for code generation in CMake
For QtQuickView/Qt Quick in Android as a View, we need a way for the
user to be able to choose QML files they would like to generate
Java classes for, to be able to access the properties of the QML
component from Java/Kotlin code.

When QT_QML_GENERATE_JAVA_CLASS property is set for a QML source
file, that file is added to the deployment json file, which
androiddeployqt will then use to generate the Java classes based
on the received values.

Task-number: QTBUG-124846
Change-Id: I393e6c1969e32fc4d9a5fca7c75f11583245b28a
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>