QtQml: Allow "using" declarations in QML type registrations
This makes qmltyperegistrar lexically replace occurrences of the alias
with the original type. It therefore assumes that the types are actually
the same. It cannot check whether this assumption holds, and you can
actually provide "using" declarations for types that are not the same if
you have the respective conversion operators in place.

Obviously this is quite dangerous. Therefore it remains internal for
now.

Task-number: QTBUG-125240
Change-Id: Idb54cb42fd1289cbd9438272ae3e9e01a30869eb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>