Tumbler: fix warnings when swapping contentItem
Referring to the contentItem property of Tumbler from within the
contentItem itself was a way to avoid using ids (which prevent deferred
execution). However, the new contentItem may not have a delegateHeight
property, so we move the declaration up to the Tumbler itself to ensure
that it's always available. Until QTBUG-11984 is implemented, we have
to use the "__" prefix to mark the property as "private" and
(hopefully) lessen the chance that it will show up in auto-completion
within users' IDEs, since it's not part of the public API (this is also
why we've historically avoided declaring properties here).

Fixes: QTBUG-119647
Pick-to: 6.5 6.7
Change-Id: Ie8646e605b0c7b1e804a6c4da4046613f958e51d
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Vladimir Belyavsky <belyavskyv@gmail.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>