ComboBox: fix popup palette inheritance, part two
0ad15c4101889a22bf9ff761e7d20ec515f46514 fixed the colors of
ComboBox's popup, but as a drive-by, tried to simplify several bindings
to palette properties. For example:

    background: Rectangle {
        color: control.palette.window
    }

became:

    background: Rectangle {
        color: palette.window
    }

I'm guessing I thought that the standard Item palette propagation would
kick in there and so the "control" part wasn't necessary. If
ComboBox.qml's popup had a line like this

    palette.window: "red"

then it would work, because the role is explicitly set. By setting a
role in QML, QQuickPaletteColorProvider::m_requestedPalette is set for
that item (Popup, in the example above). Just having

    palette: control.palette

isn't enough, because no role has been set and hence m_requestedPalette
won't be allocated and it will use the default values.

This patch restores the previous code to fix the regression.

Fixes: QTBUG-125529
Pick-to: 6.5 6.7 6.8
Change-Id: I2103b81a34d9e7f17992d0355befb8b904134db1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>