curve renderer: Fix uniform offset with multiview
When QSHADER_VIEW_COUNT is > 1, then we need to offset the next
uniform by the actual view count, even if the projection matrix
count is smaller and we therefore do no initialize all matrices.

This is a theoretical issue, but makes sense for consistency with
how the generic renderer handles this error.

Change-Id: Iccda597dac50a6c249df771814958e5f3754a870
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>