QQuickMenuItem: add implicitTextPadding and textPadding
Add the properties "implicitTextPadding" and "textPadding"
to MenuItem. They can be used by the style to ensure that
all MenuItems inside the same Menu end up left-aligned
WRT text.

Each MenuItem should set implicitTextPadding to be the
needed space from the left edge of the contentItem to
the text label. QQuickMenu will then iterate over all
the MenuItems inside the same Menu, and set textPadding
to be the maximum implicitTextPadding found.

All MenuItems should then use textPadding (which will end
up the same for all MenuItems) to position the text.

This API is meant to solve the problem that MenuItems
inside a single Menu can have different contents. Some
can be checkable, some can have an icon, some are just
plain text. And for several of our styles (e.g macOS and
Windows), we want the text to be left-aligned regardless
of that. Without this API, The checkmark inside a
checkable MenuItem would be left-aligned with the text
inside a plain MenuItem etc.

[ChangeLog][Controls][MenuItem] A MenuItem now has two
new properties (implicitTextPadding and textPadding)
that can be used for aligning the text across all
MenuItems inside a Menu.

Change-Id: I1f2248b31c63d6b9780d8fc77229a8b902362f70
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>