Android: Add QtAbstractListModel Java API
QtAbstractListModel is a standard interface for models that
represent their data as a simple non-hierarchical sequence
of items. The C++ QAbstractItemModel is its counterpart.

QtAbstractListModel extends QtAbstractItemModel just like
the C++ API. In the C++ API, the parent(), hasChildren(),
and columnCount() became private. Narrowing down access
modifiers is not allow in Java, hence instead we made them
final in the Java API to achieve nearly the same intention.

Task-number: QTBUG-125888
Change-Id: I43a25cad09d3e0bcbbf35c8beaf4546d443dbf3f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>