Enable popup windows for QtQuick.Dialogs
The dialogs in QtQuick.Dialogs look so much better on desktop systems,
when they appear inside real top level windows.

This change enables the feature for all QtQuick.Dialogs types, and
makes manual changes to each dialog, to make sure that they look good.

This includes changes that hides the title from the header, since the
title can now appear inside the window decoration instead.

Lots of tests had to be changed, as a result, to properly generate and
deliver events to the correct items.

In addition, I've added some helper functions to dialogHelper, to handle
cases we're we need to wait for a popup window to appear. Showing popups
with dedicated windows, takes longer than simply creating a new item in
an existing window.

[ChangeLog] All dialogs in QtQuick.Dialogs will now use popup windows,
if able.

Fixes: QTBUG-126006
Pick-to: 6.8
Change-Id: Ieca6ae643d4f97b1bed648e448bbcd236a50e7e7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>