Restructure builtins and QtQml.Base
The QtQml library should hold the builtins and provide no plugin. Move
the types currently exposed in QtQml.Base to QtQml where it makes sense.
Anonymous object types as well as sequence types and value types can
well be exposed as builtins. This makes everybody's life easier since
you now can universally depend on their availability.

The Qt object, despite being a named object type, also becomes a builtin
because you always have the "Qt" member of the JavaScript global object
which holds the same thing. So, formally exposing "Qt" as builtin
doesn't really add anything new.

QQmlLoggingCategory is split up into two classes, not only because we
need the base class when printing to the console from QtQml, but also
because this paves the way for compile time identification of logging
categories as first argument to the console methods.

For QQmlLocale we have to refer to a different trick. The value type and
the QQmlLocale "namespace" have to be builtins, but the "Locale" name,
due to being uppercase and versioned, has to be part of QtQml. We
transform QQmlLocale into a struct so that we can inherit the enums from
QLocale and extend a namespace with it in QtQml.

Pick-to: 6.8
Fixes: QTBUG-119473
Fixes: QTBUG-125765
Change-Id: Ica59390e54c473e33b4315f4384b724c870c1062
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>